+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "grc",
 	"displayName": "GitHub Repository Creator (GRC)",
-	"description": "Automatically create GitHub repositories using custom templates.",
+	"description": "Automatically start GitHub repositories in your workspace using custom templates.",
 	"version": "0.0.2",
 	"author": "Arthur Sudbrack Ibarra",
 	"icon": "assets/grc-logo.png",
