+++ src/configurations/user-preferences.ts
@@ -1,10 +1,10 @@
 import * as vscode from "vscode";
 
-export enum SettingsJSON {
+export enum UserPreferences {
   showAuthenticationMessages = "grc.showAuthenticationErrorMessages",
   showAuthenticationStatusBar = "grc.showAuthenticationStatusBar",
 }
 
-export function getConfig(key: SettingsJSON): any {
+export function getConfig(key: UserPreferences): any {
   return vscode.workspace.getConfiguration().get(key);
 }
