+++ src/extension.ts
@@ -14,6 +14,7 @@ import {
   getUser,
   createTemplate,
 } from "./grc/commands";
+import { showAuthMessage } from "./verifications/actions";
 import {
   checkGRCInstallation,
   checkGRCVersion,
@@ -21,12 +22,17 @@ import {
   checkRestartVSCode,
   checkIfAlreadyGitRepository,
   checkUserAthenticated,
-} from "./utils/checks";
-import {
-  getWorkingDirectory,
-  showAuthMessage,
-  updateAuthenticationStatusBar,
-} from "./utils/other";
+} from "./verifications/checks";
+import { getWorkingDirectory } from "./vscode-components/workspace";
+import { updateAuthenticationStatusBar } from "./vscode-components/status-bar";
+
+export enum ExtensionCommands {
+  installGRC = "grc.install-grc",
+  authenticate = "grc.authenticate",
+  startRepository = "grc.start-repository",
+  createTemplate = "grc.create-template",
+  addCollaborator = "grc.add-collaborator",
+}
 
 export function activate(context: vscode.ExtensionContext) {
   console.log("GRC extension activated.");
@@ -42,7 +48,7 @@ export function activate(context: vscode.ExtensionContext) {
     Command 1: Install GRC.
   */
   context.subscriptions.push(
-    vscode.commands.registerCommand("grc.install-grc", async () => {
+    vscode.commands.registerCommand(ExtensionCommands.installGRC, async () => {
       if (isGRCInstalled()) {
         vscode.window.showInformationMessage("GRC is already installed.");
         return;
@@ -83,128 +89,134 @@ export function activate(context: vscode.ExtensionContext) {
     Command 2: Authenticate.
   */
   context.subscriptions.push(
-    vscode.commands.registerCommand("grc.authenticate", async () => {
-      if (!checkRestartVSCode()) {
-        return;
+    vscode.commands.registerCommand(
+      ExtensionCommands.authenticate,
+      async () => {
+        if (!checkRestartVSCode()) {
+          return;
+        }
+        if (!checkGRCInstallation()) {
+          return;
+        }
+        if (!checkGRCVersion()) {
+          return;
+        }
+        const accessToken = await vscode.window.showInputBox({
+          prompt: "Enter your GitHub access token:",
+          password: true,
+        });
+        if (!accessToken) {
+          return;
+        }
+        const authenticated = authenticate(accessToken);
+        if (authenticated) {
+          showAuthMessage();
+          updateAuthenticationStatusBar(true, `GRC (${getUser()?.name})`);
+        } else {
+          vscode.window.showErrorMessage(
+            "(GRC) Authentication failed. Your access token is invalid."
+          );
+        }
       }
-      if (!checkGRCInstallation()) {
-        return;
-      }
-      if (!checkGRCVersion()) {
-        return;
-      }
-      const accessToken = await vscode.window.showInputBox({
-        prompt: "Enter your GitHub access token:",
-        password: true,
-      });
-      if (!accessToken) {
-        return;
-      }
-      const authenticated = authenticate(accessToken);
-      if (authenticated) {
-        showAuthMessage();
-        updateAuthenticationStatusBar(true, `GRC (${getUser()?.name})`);
-      } else {
-        vscode.window.showErrorMessage(
-          "(GRC) Authentication failed. Your access token is invalid."
-        );
-      }
-    })
+    )
   );
   /*
     Command 3: Start Repository.
   */
   context.subscriptions.push(
-    vscode.commands.registerCommand("grc.start-repository", async () => {
-      if (!checkRestartVSCode()) {
-        return;
-      }
-      if (!checkGRCInstallation()) {
-        return;
-      }
-      if (!checkGRCVersion()) {
-        return;
-      }
-      if (!checkUserAthenticated()) {
-        return;
-      }
-      if (await checkIfAlreadyGitRepository()) {
-        vscode.window.showErrorMessage(
-          "Error: Your current directory is already a git repository."
-        );
-        return;
-      }
-      const templates = getTemplates();
-      if (templates === null) {
-        vscode.window.showErrorMessage(
-          "Error: An unexpected error occurred. Try again later."
-        );
-        return;
-      }
-      if (templates.length === 0) {
-        vscode.window.showInformationMessage("You have no templates to use.");
-        return;
-      }
-      const templateName = await vscode.window.showQuickPick(templates, {
-        placeHolder: "Choose a template to use:",
-      });
-      if (!templateName) {
-        return;
-      }
-      const user = getUser();
-      const repoName = await vscode.window.showInputBox({
-        placeHolder: `(${user?.name}) Enter a name for the repository:`,
-      });
-      if (!repoName) {
-        return;
-      }
-      const repoDescription = await vscode.window.showInputBox({
-        placeHolder: "Enter a description for the repository:",
-      });
-      if (!repoDescription) {
-        return;
-      }
-      const workingDirectory = getWorkingDirectory();
-      if (!workingDirectory) {
-        vscode.window.showErrorMessage("Error: No workspace folder found.");
-        return;
-      }
-      const created = chooseTemplate(
-        templateName,
-        repoName,
-        repoDescription,
-        workingDirectory
-      );
-      if (created) {
-        const repoURL = getRepoURL(repoName);
-        if (!repoURL) {
+    vscode.commands.registerCommand(
+      ExtensionCommands.startRepository,
+      async () => {
+        if (!checkRestartVSCode()) {
+          return;
+        }
+        if (!checkGRCInstallation()) {
+          return;
+        }
+        if (!checkGRCVersion()) {
+          return;
+        }
+        if (!checkUserAthenticated()) {
+          return;
+        }
+        if (await checkIfAlreadyGitRepository()) {
           vscode.window.showErrorMessage(
-            "Error: Failed to get the repository URL."
+            "Error: Your current directory is already a git repository."
           );
           return;
         }
-        vscode.window
-          .showInformationMessage(
-            `Repository created successfully for account ${user?.username}: ${repoURL}.`,
-            "Open in Browser"
-          )
-          .then((answer) => {
-            if (answer) {
-              vscode.env.openExternal(vscode.Uri.parse(repoURL));
-            }
-          });
-      } else {
-        vscode.window.showErrorMessage(
-          `Error: Failed to create repository ${repoName}.`
+        const templates = getTemplates();
+        if (templates === null) {
+          vscode.window.showErrorMessage(
+            "Error: An unexpected error occurred. Try again later."
+          );
+          return;
+        }
+        if (templates.length === 0) {
+          vscode.window.showInformationMessage("You have no templates to use.");
+          return;
+        }
+        const templateName = await vscode.window.showQuickPick(templates, {
+          placeHolder: "Choose a template to use:",
+        });
+        if (!templateName) {
+          return;
+        }
+        const user = getUser();
+        const repoName = await vscode.window.showInputBox({
+          placeHolder: `(${user?.name}) Enter a name for the repository:`,
+        });
+        if (!repoName) {
+          return;
+        }
+        const repoDescription = await vscode.window.showInputBox({
+          placeHolder: "Enter a description for the repository:",
+        });
+        if (!repoDescription) {
+          return;
+        }
+        const workingDirectory = getWorkingDirectory();
+        if (!workingDirectory) {
+          vscode.window.showErrorMessage("Error: No workspace folder found.");
+          return;
+        }
+        const created = chooseTemplate(
+          templateName,
+          repoName,
+          repoDescription,
+          workingDirectory
         );
+        if (created) {
+          const repoURL = getRepoURL(repoName);
+          if (!repoURL) {
+            vscode.window.showErrorMessage(
+              "Error: Failed to get the repository URL."
+            );
+            return;
+          }
+          vscode.window
+            .showInformationMessage(
+              `Repository created successfully for account ${user?.username}: ${repoURL}.`,
+              "Open in Browser"
+            )
+            .then((answer) => {
+              if (answer) {
+                vscode.env.openExternal(vscode.Uri.parse(repoURL));
+              }
+            });
+        } else {
+          vscode.window.showErrorMessage(
+            `Error: Failed to create repository ${repoName}.`
+          );
+        }
       }
-    })
+    )
   );
   /*
     Command 4: Create Template.
   */
   context.subscriptions.push(
-    vscode.commands.registerCommand("grc.create-template", () => {
+    vscode.commands.registerCommand(ExtensionCommands.createTemplate, () => {
       if (!checkRestartVSCode()) {
         return;
       }
@@ -221,53 +233,56 @@ export function activate(context: vscode.ExtensionContext) {
     Command 5: Add Collaborator.
   */
   context.subscriptions.push(
-    vscode.commands.registerCommand("grc.add-collaborator", async () => {
-      if (!checkRestartVSCode()) {
-        return;
-      }
-      if (!checkGRCInstallation()) {
-        return;
-      }
-      if (!checkGRCVersion()) {
-        return;
-      }
-      if (!checkUserAthenticated()) {
-        return;
-      }
-      const repoName = await vscode.window.showInputBox({
-        placeHolder: `(${
-          getUser()?.username
-        }) Enter the name of the remote repository:`,
-      });
-      if (!repoName) {
-        return;
-      }
-      const collaboratorName = await vscode.window.showInputBox({
-        placeHolder: "Enter the name of the collaborator:",
-      });
-      if (!collaboratorName) {
-        return;
-      }
-      const permission = await vscode.window.showQuickPick(
-        availablePermissions,
-        {
-          placeHolder: `Choose the permission to give to ${collaboratorName}:`,
+    vscode.commands.registerCommand(
+      ExtensionCommands.addCollaborator,
+      async () => {
+        if (!checkRestartVSCode()) {
+          return;
         }
-      );
-      if (!permission) {
-        return;
-      }
-      const added = addCollaborator(repoName, collaboratorName, permission);
-      if (added) {
-        vscode.window.showInformationMessage(
-          `Collaborator ${collaboratorName} added to ${repoName}.`
+        if (!checkGRCInstallation()) {
+          return;
+        }
+        if (!checkGRCVersion()) {
+          return;
+        }
+        if (!checkUserAthenticated()) {
+          return;
+        }
+        const repoName = await vscode.window.showInputBox({
+          placeHolder: `(${
+            getUser()?.username
+          }) Enter the name of the remote repository:`,
+        });
+        if (!repoName) {
+          return;
+        }
+        const collaboratorName = await vscode.window.showInputBox({
+          placeHolder: "Enter the name of the collaborator:",
+        });
+        if (!collaboratorName) {
+          return;
+        }
+        const permission = await vscode.window.showQuickPick(
+          availablePermissions,
+          {
+            placeHolder: `Choose the permission to give to ${collaboratorName}:`,
+          }
         );
-      } else {
-        vscode.window.showErrorMessage(
-          `Error: Failed to add ${collaboratorName} to ${repoName}.`
-        );
+        if (!permission) {
+          return;
+        }
+        const added = addCollaborator(repoName, collaboratorName, permission);
+        if (added) {
+          vscode.window.showInformationMessage(
+            `Collaborator ${collaboratorName} added to ${repoName}.`
+          );
+        } else {
+          vscode.window.showErrorMessage(
+            `Error: Failed to add ${collaboratorName} to ${repoName}.`
+          );
+        }
       }
-    })
+    )
   );
 }
